+++ src/hooks/index.ts
@@ -1,5 +1,3 @@
-export * from './queryKeys';
-export * from './optimisticList';
 export * from './shared';
 export * from './ffe';
 export * from './proposal';